refactor(provider): collapse 6 identical SimplifiedConfig provider wrappers into subclasses
GLM, DeepSeek, Kimi, Grok, Qwen, and MiniMax provider wrappers had byte-for-byte identical provider_config() and turn_provider() method bodies — the only difference was the name default and docstring. Extracts the shared logic into SimplifiedModelProvider base class in provider/simplified.py. Each provider file is now a thin subclass that only sets the name default and keeps the provider-specific docstring. All imports, registry wiring, and test paths remain unchanged.

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
