+++ src/voidcode/provider/glm.py
@@ -2,13 +2,11 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from .config import SimplifiedProviderConfig, simplified_config_to_litellm
-from .litellm_backend import LiteLLMBackendSingleAgentProvider
-from .protocol import TurnProvider
+from .simplified import SimplifiedModelProvider
 
 
 @dataclass(frozen=True, slots=True)
-class GLMModelProvider:
+class GLMModelProvider(SimplifiedModelProvider):
     """GLM (智谱AI) Model Provider.
 
     GLM provides OpenAI-compatible API at https://open.bigmodel.cn/api/paas/v4
@@ -27,11 +25,3 @@ class GLMModelProvider:
     """
 
     name: str = "glm"
-    config: SimplifiedProviderConfig | None = None
-
-    def provider_config(self):
-        return simplified_config_to_litellm(self.name, self.config)
-
-    def turn_provider(self) -> TurnProvider:
-        adapted_config = simplified_config_to_litellm(self.name, self.config)
-        return LiteLLMBackendSingleAgentProvider(name=self.name, config=adapted_config)
