+++ src/voidcode/provider/kimi.py
@@ -2,13 +2,11 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from .config import SimplifiedProviderConfig, simplified_config_to_litellm
-from .litellm_backend import LiteLLMBackendSingleAgentProvider
-from .protocol import TurnProvider
+from .simplified import SimplifiedModelProvider
 
 
 @dataclass(frozen=True, slots=True)
-class KimiModelProvider:
+class KimiModelProvider(SimplifiedModelProvider):
     """Kimi (Moonshot AI) Model Provider.
 
     Kimi provides OpenAI-compatible API at https://api.moonshot.cn/v1 (China)
@@ -26,11 +24,3 @@ class KimiModelProvider:
     """
 
     name: str = "kimi"
-    config: SimplifiedProviderConfig | None = None
-
-    def provider_config(self):
-        return simplified_config_to_litellm(self.name, self.config)
-
-    def turn_provider(self) -> TurnProvider:
-        adapted_config = simplified_config_to_litellm(self.name, self.config)
-        return LiteLLMBackendSingleAgentProvider(name=self.name, config=adapted_config)
