+++ src/voidcode/provider/minimax.py
@@ -2,13 +2,11 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from .config import SimplifiedProviderConfig, simplified_config_to_litellm
-from .litellm_backend import LiteLLMBackendSingleAgentProvider
-from .protocol import TurnProvider
+from .simplified import SimplifiedModelProvider
 
 
 @dataclass(frozen=True, slots=True)
-class MiniMaxModelProvider:
+class MiniMaxModelProvider(SimplifiedModelProvider):
     """MiniMax AI Model Provider.
 
     MiniMax provides OpenAI-compatible and Anthropic-compatible APIs.
@@ -27,11 +25,3 @@ class MiniMaxModelProvider:
     """
 
     name: str = "minimax"
-    config: SimplifiedProviderConfig | None = None
-
-    def provider_config(self):
-        return simplified_config_to_litellm(self.name, self.config)
-
-    def turn_provider(self) -> TurnProvider:
-        adapted_config = simplified_config_to_litellm(self.name, self.config)
-        return LiteLLMBackendSingleAgentProvider(name=self.name, config=adapted_config)
