+++ src/voidcode/provider/qwen.py
@@ -2,13 +2,11 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from .config import SimplifiedProviderConfig, simplified_config_to_litellm
-from .litellm_backend import LiteLLMBackendSingleAgentProvider
-from .protocol import TurnProvider
+from .simplified import SimplifiedModelProvider
 
 
 @dataclass(frozen=True, slots=True)
-class QwenModelProvider:
+class QwenModelProvider(SimplifiedModelProvider):
     """Qwen (通义千问) Model Provider.
 
     Qwen provides OpenAI-compatible API at https://dashscope.aliyuncs.com/compatible-mode/v1
@@ -25,11 +23,3 @@ class QwenModelProvider:
     """
 
     name: str = "qwen"
-    config: SimplifiedProviderConfig | None = None
-
-    def provider_config(self):
-        return simplified_config_to_litellm(self.name, self.config)
-
-    def turn_provider(self) -> TurnProvider:
-        adapted_config = simplified_config_to_litellm(self.name, self.config)
-        return LiteLLMBackendSingleAgentProvider(name=self.name, config=adapted_config)
