+++ src/voidcode/provider/simplified.py
@@ -1,0 +1,20 @@
+from __future__ import annotations
+
+from dataclasses import dataclass
+
+from .config import SimplifiedProviderConfig, simplified_config_to_litellm
+from .litellm_backend import LiteLLMBackendSingleAgentProvider
+from .protocol import TurnProvider
+
+
+@dataclass(frozen=True, slots=True)
+class SimplifiedModelProvider:
+    name: str
+    config: SimplifiedProviderConfig | None = None
+
+    def provider_config(self):
+        return simplified_config_to_litellm(self.name, self.config)
+
+    def turn_provider(self) -> TurnProvider:
+        adapted_config = simplified_config_to_litellm(self.name, self.config)
+        return LiteLLMBackendSingleAgentProvider(name=self.name, config=adapted_config)
